Make node table timers shared_ptr
This is required to keep timers alive if handlers are executed after the
node table has been destroyed. This shouldn't happen in practice since
the node table lifetime is tied to the host lifetime and if the host is
destroyed then the ioservice is no longer running, but these changes will
help avoid use-after-frees if the node table/host relationship changes
in the future.

Also have the node table inherit enable_shared_from_this so that the
node table can be kept alive in ping handlers. Similar to the timers
case, this isn't needed but is good design in case the node table/host
relationship changes in the future. Note that this change requires that
the code which starts the discovery and eviction processes is moved to a
separate function (NodeTable::start) because we cannot call shared_from_this() before the
object is fully constructed.