Fix unreliable network unit test
Network test findNodeIsSentAfterPong is unreliable because it expects the first node table to send a FindNode packet to the second node table before the second node table sends a FindNode packet to the first node table. However, there's no guarantee that this will happen because both node tables are running discovery code in separate threads. While starting the first node table before the second node table results in the expected behavior most of the time, occasionally the second node table will send the FindNode packet first resulting in the test failing because the second node table receives a Neighbours packet when it expects a FindNode packet.

I've fixed this by replacing one of the node tables with a TestUDPSocketHost - since there's only 1 discovery process running now we can deterministically know when to expect a FindNode packet to be sent by the node table.