Remove shared_ptr node table timers
shared_ptr was used to pass an instance of a timer to its handler to ensure its kept alive until the handler is executed. However, this is unnecessary since the node table inherits from enable_shared_from_this and the timer handlers are now lambdas which take a shared this ptr.